fixed most compiler warnings, code cleanup
svn path=/trunk/mono/; revision=2250


Commit migrated from https://github.com/mono/mono/commit/bab65ef9835dab7c47a23aa8cdd8acc4eaad8d92

+++ src/mono/mono/jit/emit-x86.c
@@ -1205,7 +1205,7 @@ arch_handle_exception (struct sigcontext *ctx, gpointer obj)
 		ctx->ebx = lmf->ebx;
 		ctx->ebp = lmf->ebp;
 		ctx->eip = lmf->eip;
-		ctx->esp = lmf;
+		ctx->esp = (unsigned long)lmf;
 
 		if (mono_object_isinst (obj, mono_defaults.exception_class)) {
 			char  *strace = mono_string_to_utf8 (((MonoException*)obj)->stack_trace);
